packages/justhtml/justhtml-0.12.0.tar.gz/justhtml-0.12.0/tests/justhtml-tests/branch_coverage.dat

#data
<math><mi>\x00</mi></math>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>


#data
<math><mi>\x00\x0c</mi></math>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>


#data
<math><mi>   </mi></math>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "   "


#data
</template>
#errors
#document
| <html>
|   <head>
|   <body>


#data
<body><input type="hidden">
#errors
#document
| <html>
|   <head>
|   <body>
|     <input>
|       type="hidden"


#data
<html><body><html attr="test">
#errors
#document
| <html>
|   attr="test"
|   <head>
|   <body>


#data
<select><optgroup><option>1</option></optgroup></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <optgroup>
|         <option>
|           "1"


#data
<select><option>1</option></optgroup></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         "1"


#data
<!DOCTYPE html><html><head></head><frameset></frameset> 
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
|   " "


#data
<!DOCTYPE html><html><head></head><frameset></frameset>abc
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>


#data
<table><input id="x" class="y" type="hidden"></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <input>
|         class="y"
|         id="x"
|         type="hidden"


#data
<math><annotation-xml encoding="text/plain"><div></div></annotation-xml></math>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/plain"
|     <div>


#data
<math><annotation-xml><div></div></annotation-xml></math>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|     <div>


#data
<svg><font color="red">text</font></svg>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <font>
|       color="red"
|       "text"


#data
<select><selectedcontent></selectedcontent><option>A</option><option selected>B</option></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <selectedcontent>
|         "B"
|       <option>
|         "A"
|       <option>
|         selected=""
|         "B"


#data
<select><selectedcontent></selectedcontent><option>A</option><option>B</option></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <selectedcontent>
|         "A"
|       <option>
|         "A"
|       <option>
|         "B"


#data
<b><select></b></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <select>


#data
<a href="x"><select></a></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <a>
|       href="x"
|       <select>


#data
<select><plaintext></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <plaintext>
|         "</select>"


#data
<svg><g><rect></rect><circle></circle><path></path></g></svg><div></div>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         <svg rect>
|         <svg circle>
|         <svg path>
|     <div>


#data
<math><annotation-xml encoding="text/html"><div><table><tr></tr></table></div></annotation-xml></math>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/html"
|         <div>
|           <table>
|             <tbody>
|               <tr>


#data
<select></a></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>


#data
<select></b></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>


#data
<math><annotation-xml foo="bar"><div>test</div></annotation-xml></math>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         foo="bar"
|     <div>
|       "test"


#data
<svg><g><rect><circle><path><div>text</div></path></circle></rect></g></svg>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>
|         <svg rect>
|           <svg circle>
|             <svg path>
|     <div>
|       "text"


#data
<select><aside>text</aside></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>
|       "text"


#data
<select><selectedcontent></selectedcontent><option disabled>A</option></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <selectedcontent>
|         "A"
|       <option>
|         disabled=""
|         "A"


#data
<select><selectedcontent></selectedcontent><option class="x">A</option><option selected>B</option></select>
#errors
#document
| <html>
|   <head>
|   <body>
|     <select>
|       <selectedcontent>
|         "B"
|       <option>
|         class="x"
|         "A"
|       <option>
|         selected=""
|         "B"
